Separate out operator vs model testing. - create BaseTester class and derived OpTester/ModelTester classes to limit APIs to what is applicable - cleanup structure   - don't expose member variables either directly or via public methods returning them   - split out checkers so they can be easily re-used   - refactor the GradientOpTester usage     - it required a lot of OpTester externals to be exposed and no other tests used this     - it also returned Status through various parts which prevented the usage of the google test macros which provide better output. change to return void - fix some other minor issues
