+++ onnxruntime/core/platform/windows/debug_alloc.cc
@@ -197,7 +197,7 @@ Memory_LeakCheck::~Memory_LeakCheck() {
 
     std::string string;
     char buffer[1024];
-    _snprintf_s(buffer, _TRUNCATE, "%IX bytes at location 0x%08IX\n", entry.cbData - sizeof(MemoryBlock), UINT_PTR(pBlock));
+    _snprintf_s(buffer, _TRUNCATE, "%Id bytes at location 0x%08IX\n", entry.cbData - sizeof(MemoryBlock), UINT_PTR(pBlock));
     string.append(buffer);
     for (auto& p : block.m_pTraces) {
       if (!p) break;
