+++ onnxruntime/test/framework/inference_session_test.cc
@@ -57,6 +57,7 @@
 #include "test/util/include/inference_session_wrapper.h"
 
 #include "gtest/gtest.h"
+#include "gmock/gmock.h"
 
 using namespace std;
 using namespace ONNX_NAMESPACE;
@@ -2017,7 +2018,8 @@ TEST(InferenceSessionTests, TestStrictShapeInference) {
   // we also need for the output to be valid so OpTester doesn't throw so add an Unsqueeze after the Shape.
   class OpTesterWithReshape : public OpTester {
    public:
-    OpTesterWithReshape() : OpTester("Shape", 7) {
+    // don't cache the model as we're testing failure during Graph::Resolve
+    OpTesterWithReshape() : OpTester("Shape", 7, kOnnxDomain, /* cache_model */ false) {
     }
 
    protected:
