+++ onnxruntime/test/providers/cpu/tensor/cast_op_test.cc
@@ -54,7 +54,7 @@ template <typename SrcType,
           typename DstType>
 void TestCastOp(gsl::span<const SrcType> input,
                 gsl::span<const DstType> output,
-                const std::vector<int64_t>& dimensions,
+                const BaseTester::DimsVariant& dimensions,
                 OpTester::ExpectResult expect_result = OpTester::ExpectResult::kExpectSuccess,
                 const std::string& expected_failure_string = "",
                 int opset = 13,
@@ -136,7 +136,7 @@ struct CastNonStringTester {
     auto output_span = gsl::make_span<DstType>(output_buffer.get(), size);
     CastSpan<SrcType, DstType>(input_span, output_span);
 
-    TestCastOp<SrcType, DstType>(input_span, output_span, GetShapeVector(shape));
+    TestCastOp<SrcType, DstType>(input_span, output_span, shape.AsShapeVector());
   }
 };
