+++ orttraining/orttraining/test/training_ops/cpu/nn/dropout_op_test.cc
@@ -79,7 +79,7 @@ void RunDropoutTest(const bool use_mask, const std::vector<int64_t>& input_shape
 
   auto output_verifier = [&](const std::vector<OrtValue>& fetches, const std::string& provider_type) {
     ASSERT_GE(fetches.size(), 1);
-    const auto& output_tensor = FetchTensor(fetches[0]);
+    const auto& output_tensor = fetches[0].Get<Tensor>();
     auto output_span = output_tensor.DataAsSpan<float>();
 
     const auto num_dropped_values = std::count(output_span.begin(), output_span.end(), 0.0f);
@@ -100,7 +100,7 @@ void RunDropoutTest(const bool use_mask, const std::vector<int64_t>& input_shape
 
     if (use_mask) {
       ASSERT_GE(fetches.size(), 2);
-      const auto& mask_tensor = FetchTensor(fetches[1]);
+      const auto& mask_tensor = fetches[1].Get<Tensor>();
       auto mask_span = mask_tensor.DataAsSpan<bool>();
       ASSERT_EQ(mask_span.size(), output_span.size()) << "provider: " << provider_type;
