+++ orttraining/orttraining/test/training_ops/function_op_test_utils.h
@@ -8,13 +8,13 @@ namespace onnxruntime {
 namespace test {
 using ONNX_NAMESPACE::AttributeProto;
 
-typedef std::vector<std::vector<float>> TwoDArray;
+using TwoDArray = std::vector<std::vector<float>>;
 
 class OpFunctionTester : public OpTester {
  public:
   OpFunctionTester(const char* op, int opset_version = 9, const char* domain = onnxruntime::kOnnxDomain)
       : OpTester(op, opset_version, domain) {}
-  TwoDArray RunFunctionBodyGraphOnCPU();
+  void RunFunctionBodyGraphOnCPU(TwoDArray& results);
   virtual ~OpFunctionTester();
 };
